ai/agent/src/pyproject.toml
[project]
name = "city-expert-agent"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "acp-sdk>=0.9.1",
    "crewai[tools]==0.120.1",
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = ["ruff>=0.11.10"]
